# bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "bbva-backend, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Aforo": {
      "post": {
        "tags": [
          "Aforo"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AforoData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Agencias": {
      "get": {
        "tags": [
          "Agencias"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Agencias/all": {
      "get": {
        "tags": [
          "Agencias"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CamarasAgencias/GetImagenAgencia": {
      "get": {
        "tags": [
          "CamarasAgencias"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idAgencia",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cliente": {
      "post": {
        "tags": [
          "Cliente"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "dni",
            "type": "string"
          },
          {
            "in": "query",
            "name": "password",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tickets/CrearTicket": {
      "post": {
        "tags": [
          "Tickets"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "dni",
            "type": "string"
          },
          {
            "in": "query",
            "name": "idAgencia",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tickets/AtenderTicket": {
      "post": {
        "tags": [
          "Tickets"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idTicket",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tickets/TerminarTicket": {
      "post": {
        "tags": [
          "Tickets"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idTicket",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tickets/ObtenerTicketsAtendiendose": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idAgencia",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tickets/ObtenerTicketsCreadosSinAtender": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idAgencia",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tickets/ListarTicketsHoy": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tickets/ListarTickets": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "AforoData": {
      "type": "object",
      "properties": {
        "idAgencia": {
          "format": "int32",
          "type": "integer"
        },
        "aforo": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}